feat: create transaction list component
